Feat: MessageContainer를 만들어 Message 컴포넌트 동적으로 생성
